feat: implement SourceAdapter pattern + SourceRegistry (closes #1)
Refactor discovery layer from monolithic content.ts/community.ts into
pluggable adapter pattern:

- SourceAdapter interface (inspired by Squad SDK SkillSource pattern)
- SourceRegistry with parallel execution via Promise.allSettled()
- RSSSourceAdapter extracted from content.ts
- GitHubSourceAdapter extracted from community.ts
- Shared helpers consolidated (generateCanonicalId, keywords, etc.)
- Graceful degradation: failed adapters logged, others continue

New adapters can be added by implementing SourceAdapter and registering
with the registry — no changes to orchestrator needed.

Closes #1
