+++ .squad/decisions/inbox/copilot-directive-2026-03-12T09-17.md
@@ -1,0 +1,9 @@
+### 2026-03-12T09:17: User directives
+**By:** Brady (via Copilot)
+**What:**
+1. Use the Squad SDK for all new functionality — the adapter/source system must be pluggable using the Squad SDK.
+2. File issues on the repo as PRDs and implementation plans before doing work.
+3. Work on branches and PR back to main.
+4. Team reviews PRs before merge — quality must meet Beth's standards.
+5. Don't stop until Beth has a useful system covering all sources she needs.
+**Why:** User request — captured for team memory
